+++ client/src/components/document/DocumentContents.tsx
