+++ client/src/components/document/DocumentFooter.tsx
