+++ client/src/components/document/DocumentHeader.tsx
@@ -1,17 +1,18 @@
 import React from 'react';
-import { useQueryClient } from '@tanstack/react-query';
-import URLS from '@constants/urls';
 import { useNavigate } from 'react-router-dom';
+import URLS from '@constants/urls';
 import { WikiDocument } from '@type/DocumentType';
-import Button from './Button';
+import { useQueryClient } from '@tanstack/react-query';
+import DocumentTitle from '../common/DocumentTitle';
+import Button from '../common/Button';
 
-interface MobileDocumentHeaderProps {
-  docs: WikiDocument;
+interface DocumentHeaderProps {
+  wiki: WikiDocument;
 }
 
-const MobileDocumentHeader = ({ docs }: MobileDocumentHeaderProps) => {
-  const queryClient = useQueryClient();
+const DocumentHeader = ({ wiki }: DocumentHeaderProps) => {
   const navigate = useNavigate();
+  const queryClient = useQueryClient();
 
   const refreshData = () => {
     queryClient.removeQueries();
@@ -23,23 +24,24 @@ const MobileDocumentHeader = ({ docs }: MobileDocumentHeaderProps) => {
   };
 
   const goEditPage = () => {
-    navigate(URLS.EDIT, { state: docs });
+    navigate(URLS.EDIT, { state: wiki });
   };
 
   const goLogsPage = () => {
-    navigate(`${URLS.WIKI}/${docs.title}/${URLS.LOGS}`, { state: docs });
+    navigate(`${URLS.WIKI}/${wiki.title}/${URLS.LOGS}`, { state: wiki });
   };
 
   return (
-    <div className="md:hidden">
-      <fieldset className="flex gap-2 max-md:w-full max-md:justify-center">
+    <header className="max-md:flex-col-reverse max-md:gap-4 flex justify-between w-full">
+      <DocumentTitle title={wiki.title} />
+      <fieldset className="flex gap-2 max-md:hidden">
         <Button style="tertiary" size="xs" text="새로고침" onClick={refreshData} />
         <Button style="tertiary" size="xs" text="편집하기" onClick={goEditPage} />
         <Button style="tertiary" size="xs" text="편집로그" onClick={goLogsPage} />
         <Button style="primary" size="xs" text="작성하기" onClick={goPostPage} />
       </fieldset>
-    </div>
+    </header>
   );
 };
 
-export default MobileDocumentHeader;
+export default DocumentHeader;
