+++ client/src/components/logs/LogsContents.tsx
