+++ client/src/components/write/TitleInputField.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Input from './Input';
+import Input from '../common/Input';
 
 interface TitleInputFieldProps {
   titleState: {
