+++ client/src/components/write/WritePage.tsx
